requirements.txt
# Core Dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1
pydub>=0.25.1

# DeepFilterNet for noise reduction
deepfilternet>=0.5.6

# Voice Activity Detection
webrtcvad>=2.0.10
silero-vad>=4.0.0

# Diarization
pyannote.audio>=3.1.0

# Speech Recognition
openai-whisper>=20231117

# Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
